+++ react/components/NestedExtensionPoints.tsx
@@ -17,8 +17,9 @@ interface Props {
 
 export default class NestedExtensionPoints extends PureComponent<Props> {
   public static propTypes = {
+    breakPoint: PropTypes.object,
     page: PropTypes.string.isRequired,
-    query: PropTypes.object,
+    query: PropTypes.object
   }
 
   public getPageParams(runtime: RenderContext, name: string) {
@@ -38,20 +39,7 @@ export default class NestedExtensionPoints extends PureComponent<Props> {
     const getNestedExtensionPoints = (runtime: RenderContext) => {
       return reverse.reduce((acc: JSX.Element | null, value: string, index: number) => {
         const { breakPoint } = this.props
-        if (breakPoint && breakPoint.point && breakPoint.point === value) {
-          if (breakPoint.loading) {
-            return <div className="flex justify-center ma4"><Loading /></div>
-          } else if (breakPoint.logged) {
-            return <ExtensionPoint
-              id={value}
-              query={query}
-              params={this.getPageParams(runtime, segments.slice(0, segments.length - index).join('/'))}>
-              {acc}
-            </ExtensionPoint>
-          }
-          return null
-        }
-        return (
+        const extensionPoint = (
           <ExtensionPoint
             id={value}
             query={query}
@@ -59,6 +47,15 @@ export default class NestedExtensionPoints extends PureComponent<Props> {
             {acc}
           </ExtensionPoint>
         )
+        if (breakPoint && breakPoint.point && breakPoint.point === value) {
+          if (breakPoint.loading) {
+            return <div className="flex justify-center ma4"><Loading /></div>
+          } else if (breakPoint.logged) {
+            return extensionPoint
+          }
+          return null
+        }
+        return extensionPoint
       }, null as JSX.Element | null)
     }