+++ CHANGELOG.md
@@ -7,11 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- Auth treatment and redirectioning of the store
+- Auth treatment and redirect to the login page.
 
 ## [7.11.3] - 2018-6-21
 ### Removed
-- Remove support for root extension without component
+- Remove support for root extension without component.
 
 ## [7.11.2] - 2018-6-21
 ### Fixed