+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Children with missing props in the `MaybeAuth` component.
 
+## [7.12.6] - 2018-07-10
+
 ## [7.12.0] - 2018-07-09
 ### Added
 - Auth treatment and redirect to the login page.